Removed code duplication in test framework; moved Stopwatch to host
