+++ rustzx-utils/src/lib.rs
@@ -1,0 +1,8 @@
+#![no_std]
+
+#[cfg(feature = "std")]
+extern crate std;
+
+pub mod palette;
+#[cfg(feature = "std")]
+pub mod stopwatch;
