+++ rustzx/Cargo.toml
@@ -14,6 +14,7 @@ edition = "2018"
 [dependencies]
 sdl2 = { version = "0.34", features = ["unsafe_textures", "bundled", "static-link"] }
 rustzx-core = { version = "0.13", path =  "../rustzx-core", features = ["full"] }
+rustzx-utils = { version = "0.1", path =  "../rustzx-utils", features = ["std"] }
 log = "0.4"
 anyhow = "1"
 structopt = "0.3"
