+++ rustzx/src/app/video/palette.rs
@@ -1,4 +1,5 @@
 use rustzx_core::zx::video::colors::{ZXBrightness, ZXColor};
+use rustzx_utils::palette::rgba::ORIGINAL as DEFAULT_PALETTE;
 
 type ColorRgba = [u8; 4];
 
@@ -11,26 +12,7 @@ pub struct Palette {
 impl Default for Palette {
     fn default() -> Self {
         Palette {
-            colors: [
-                // normal
-                0x000000FF_u32.to_be_bytes(),
-                0x0000CDFF_u32.to_be_bytes(),
-                0xCD0000FF_u32.to_be_bytes(),
-                0xCD00CDFF_u32.to_be_bytes(),
-                0x00CD00FF_u32.to_be_bytes(),
-                0x00CDCDFF_u32.to_be_bytes(),
-                0xCDCD00FF_u32.to_be_bytes(),
-                0xCDCDCDFF_u32.to_be_bytes(),
-                // bright
-                0x000000FF_u32.to_be_bytes(),
-                0x0000FFFF_u32.to_be_bytes(),
-                0xFF0000FF_u32.to_be_bytes(),
-                0xFF00FFFF_u32.to_be_bytes(),
-                0x00FF00FF_u32.to_be_bytes(),
-                0x00FFFFFF_u32.to_be_bytes(),
-                0xFFFF00FF_u32.to_be_bytes(),
-                0xFFFFFFFF_u32.to_be_bytes(),
-            ],
+            colors: DEFAULT_PALETTE,
         }
     }
 }
