+++ rustzx/src/host/mod.rs
@@ -8,6 +8,7 @@ use rustzx_core::{
     host::{FrameBuffer, Host, HostContext, RomFormat, RomSet, Screen, Snapshot, Tape},
     zx::machine::ZXMachine,
 };
+use rustzx_utils::stopwatch::InstantStopwatch;
 use std::{collections::VecDeque, fs::File, path::Path};
 
 const SUPPORTED_SNAPSHOT_FORMATS: [&str; 1] = ["sna"];
@@ -18,6 +19,7 @@ pub struct AppHost;
 
 impl Host for AppHost {
     type Context = AppHostContext;
+    type EmulationStopwatch = InstantStopwatch;
     type FrameBuffer = RgbaFrameBuffer;
     type TapeAsset = FileAsset;
 }
